+++ src/utils/clack-utils.ts
@@ -902,7 +902,7 @@ export async function getOrAskForProjectData(
       platform: platform,
       orgSlug: options.orgSlug,
       projectSlug: options.projectSlug,
-      partner: options.partner,
+      comingFrom: options.comingFrom,
     }),
   );
 
@@ -1045,9 +1045,10 @@ export async function askForWizardLogin(options: {
     | 'flutter';
   orgSlug?: string;
   projectSlug?: string;
-  partner?: string;
+  comingFrom?: string;
 }): Promise<WizardProjectData> {
-  const { orgSlug, projectSlug, url, platform, promoCode, partner } = options;
+  const { orgSlug, projectSlug, url, platform, promoCode, comingFrom } =
+    options;
 
   Sentry.setTag('has-promo-code', !!promoCode);
 
@@ -1081,9 +1082,9 @@ export async function askForWizardLogin(options: {
   if (promoCode) {
     loginUrl.searchParams.set('code', promoCode);
   }
-
-  if (partner) {
-    loginUrl.searchParams.set('partner', partner);
+  
+  if (comingFrom) {
+    loginUrl.searchParams.set('partner', comingFrom);
   }
 
   const urlToOpen = loginUrl.toString();