+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- feat: Add partner parameter ([#837](https://github.com/getsentry/sentry-wizard/pull/837))
+- feat: Add `coming-from` parameter ([#837](https://github.com/getsentry/sentry-wizard/pull/837))
 
 ## 4.1.0